pot distribution

+++ src/inout.rs
@@ -19,13 +19,13 @@ pub fn print_cards(deck: &cards::Deck) {
     }
 }
 
-pub fn print_combination(comb: &[usize]) {
+pub fn print_combination(comb: &[u32]) {
     let combinations_names = ["Carte haute", "Paire", "Deux paires", "Brelan", "Quinte", "Couleur", "Full", "Carré", "Quinte couleur", "Quinte royale"];
     let values_names = ["Un", "Deux", "Trois", "Quatre", "Cinq", "Six", "Sept", "Huit", "Neuf", "Dix", "Valet", "Dame", "Roi", "As"];
-    println!("{}", combinations_names[comb[0]-1]);
+    println!("{}", combinations_names[(comb[0]-1) as usize]);
     let mut i = 1;
     while i < 6 && comb[i] != 0 {
-        print!("{} ", values_names[comb[i]-1]);
+        print!("{} ", values_names[(comb[i]-1) as usize]);
         i += 1;
     }
     println!("");
